事前準備のデータ挿入用実装
