+++ src/main/java/com/example/controller/CreateTableController.java
@@ -4,27 +4,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.example.service.OriginalService;
+import com.example.service.CreateTableService;
 
 @Controller
 @RequestMapping("/category")
-public class CategoryController {
+public class CreateTableController {
 	
 	@Autowired
-	private OriginalService originalService;
+	private CreateTableService createTableService;
 	
 	@RequestMapping("/largeCategoryRegister")
 	public void largeCategoryRegister() {
-		originalService.largeCategoryRegister();
+		createTableService.largeCategoryRegister();
 	}
 	
 	@RequestMapping("/middleCategoryRegister")
 	public void middleCategoryRegister() {
-		originalService.middleCategoryRegister();
+		createTableService.middleCategoryRegister();
 	}
 	
 	@RequestMapping("/smallCategoryRegister")
 	public void smallCategoryRegister() {
-		originalService.smallCategoryRegister();
+		createTableService.smallCategoryRegister();
 	}
 }
