+++ src/main/resources/com/example/mapper/createTable/CategoryMapper.xml
@@ -3,7 +3,7 @@
 			PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
         	
-<mapper namespace="com.example.mapper.CategoryMapper">
+<mapper namespace="com.example.mapper.createTable.CategoryMapper">
 	
 	<!-- DB登録 -->
 	<insert id="insert">
