+++ src/main/resources/com/example/mapper/createTable/OriginalMapper.xml
@@ -3,7 +3,7 @@
 			PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
         	
-<mapper namespace="com.example.mapper.OriginalMapper">
+<mapper namespace="com.example.mapper.createTable.OriginalMapper">
 	
 	<!-- 全件選択 -->
 	<select id="findAll" resultMap="originalCategory">
