+++ src/main/resources/document/sql.txt
@@ -1,12 +1,3 @@
-INSERT INTO original 
-	(name, item_condition_id, category_name, brand_name, price, shipping, item_description)
-VALUES
-	
-	
-COPY original
-	(name, condition_id, category_name, brand, price, shipping, description)
-FROM 
-	'/Users/yoshiokawakubo/var/lib/postgresql/data/DataManagementSystem/train.tsv';
 
 --  外部ファイルをインポート --
 COPY original
@@ -29,9 +20,57 @@ CREATE TABLE categories
 	, name_all text
 );
 
-DELETE FROM categories WHERE id > 10;
-
-SELECT category_name FROM original WHERE category_name LIKE '%Electronics%';
-
 --  handmadeカテゴリ > smallcategory > parent_id SQL文 --
 UPDATE categories SET parent_Id = 10 WHERE id > 1142;
+
+--  itemsテーブルの作成 --
+CREATE TABLE items (
+	id serial primary key
+	, name text
+	, condition integer
+	, category integer
+	, brand text
+	, price decimal
+	, shipping integer
+	, description text
+);
+
+--  itemsテーブルのDB挿入 --
+【1.originalでcategory_nameがnullじゃないものを挿入】
+insert into items (
+	name
+	, condition
+	, category
+	, brand
+	, price
+	, shipping
+	, description
+) 
+SELECT 
+	o.name
+	, o.item_condition_id
+	, c.id
+	, o.brand_name
+	, o.price
+	, o.shipping
+	, o.item_description
+FROM original AS o
+JOIN categories AS c ON c.name_all = o.category_name;
+【2.originalでcategory_nameがnullのものを挿入】
+insert into items (
+	name
+	, condition
+	, brand
+	, price
+	, shipping
+	, description
+) 
+SELECT 
+	o.name
+	, o.item_condition_id
+	, o.brand_name
+	, o.price
+	, o.shipping
+	, o.item_description
+FROM original AS o
+WHERE o.category_name is NULL;
